+++ doc_source/contents/developer/changelog.md
@@ -12,6 +12,7 @@
 - Add the `autoprocess` subcommand, which requires moving some subcommand arguments
   to gmprocess; this includes `eventid`, `textfile`, `label`, `num-processes`, and `overwrite`.
 - Note that moving the `label` argument to gmrecords from the subcommands means that the short flag `-l` conflicts with `log` so the short flag for log has been removed.
+- Include "passed" or "failed" for each station in export_failure_tables in addition to failure reason.
 
 ## 1.2.1 / 2022-10-04